[QOL-7740] split CLI module into Paster and Click
